Add files via upload
This is my first attempt at creating a website which is dedicated for farmers and agriculturist around India, who can access the power of data science and computer vision to do precision agriculture and increase crop yield. The aim of the website was to be very simply and user-friendly. I would appreciate any insights to improve this website.